+++ src/lib.rs
@@ -20,7 +20,6 @@
 #![doc = include_str!("../README.md")]
 
 pub mod builder;
-pub mod crypto;
 pub mod error;
 pub mod event;
 pub mod interface;
@@ -29,7 +28,6 @@ mod interfaces;
 mod mock;
 pub mod prelude;
 pub mod properties;
-pub mod registration;
 mod retry;
 mod shared;
 pub mod store;
@@ -55,14 +53,14 @@ use tokio::sync::{mpsc, RwLock};
 use transport::Disconnect;
 
 /// Re-exported internal structs
+pub use crate::error::Error;
 pub use crate::event::FromEvent;
 pub use crate::interface::Interface;
 
-use crate::error::Error;
 use crate::interface::mapping::path::MappingPath;
 use crate::interface::reference::MappingRef;
 use crate::interface::reference::PropertyRef;
-use crate::interface::{Aggregation as InterfaceAggregation, InterfaceError};
+use crate::interface::Aggregation as InterfaceAggregation;
 use crate::interfaces::Interfaces;
 use crate::shared::SharedDevice;
 use crate::store::wrapper::StoreWrapper;
@@ -199,7 +197,9 @@ impl<S, C> AstarteDeviceSdk<S, C> {
         let interfaces = self.interfaces.read().await;
         let interface = interfaces.get(interface).ok_or_else(|| {
             warn!("publish on missing interface {interface} ({path})");
-            Error::MissingInterface(interface.to_string())
+            Error::InterfaceNotFound {
+                name: interface.to_string(),
+            }
         })?;
 
         let (data, timestamp) = match interface.aggregation() {
@@ -231,7 +231,7 @@ impl<S, C> AstarteDeviceSdk<S, C> {
     {
         let mapping = interface
             .as_mapping_ref(path)
-            .ok_or_else(|| Error::MissingMapping {
+            .ok_or_else(|| Error::MappingNotFound {
                 interface: interface.interface_name().to_string(),
                 mapping: path.to_string(),
             })?;
@@ -422,7 +422,9 @@ impl<S, C> AstarteDeviceSdk<S, C> {
         let interfaces = self.interfaces.read().await;
         let interface = interfaces
             .get(interface_name)
-            .ok_or_else(|| Error::MissingInterface(interface_name.to_string()))?;
+            .ok_or_else(|| Error::InterfaceNotFound {
+                name: interface_name.to_string(),
+            })?;
 
         let object = interface
             .as_object_ref()
@@ -802,10 +804,8 @@ where
             .write()
             .await
             .remove(interface_name)
-            .ok_or_else(|| {
-                Error::Interface(InterfaceError::InterfaceNotFound {
-                    name: interface_name.to_string(),
-                })
+            .ok_or_else(|| Error::InterfaceNotFound {
+                name: interface_name.to_string(),
             })?;
 
         {